修: rotate字段类型int→float64; table只显示实际替换的isNew行
- model/types.go: Rotate int→float64，兼容JSON中浮点值
- cmd/replace.go: isNew表格行从替换成功后收集改为预先收集，
  只显示图片实际替换成功的灯位，避免空数据入表

+++ model/types.go
@@ -205,7 +205,7 @@ type ImageInfo struct {
 	} `json:"originalTransform"`
 	PageHeight float64 `json:"pageHeight"`
 	PageWidth  float64 `json:"pageWidth"`
-	Rotate     int     `json:"rotate"`
+	Rotate     float64 `json:"rotate"`
 	Width      float64 `json:"width"`
 	X          float64 `json:"x"`
 	Y          float64 `json:"y"`
